backend/requirements.txt
fastapi==0.115.1
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.22
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-dotenv==1.0.0
email-validator==2.3.0
